[8단계-refactor] MessageGenerator 클래스 단위 분리
